requirements.txt
amqp==5.2.0
aniso8601==9.0.1
Babel==2.12.1
bcrypt==4.0.1
billiard==4.2.0
blinker==1.6.2
boto3==1.28.53
botocore==1.31.53
cachelib==0.9.0
celery==5.3.5
cffi==1.15.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.4
dnspython==2.4.2
email-validator==2.0.0.post2
Flask==2.2.5
Flask-BabelEx==0.9.4
Flask-Bcrypt==1.0.1
Flask-Caching==2.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Reuploaded==1.4.0
Flask-Security-Too==5.1.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.3.3
MarkupSafe==2.1.3
packaging==23.1
passlib==1.7.4
Pillow==10.1.0
pluggy==1.3.0
prompt-toolkit==3.0.41
pycparser==2.21
pycryptodome==3.19.0
PyJWT==2.8.0
pytest==7.4.2
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
reportlab==4.0.7
s3transfer==0.6.2
six==1.16.0
speaklater==1.3
SQLAlchemy==2.0.21
tabulate==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
vine==5.1.0
wcwidth==0.2.10
Werkzeug==2.3.7
WTForms==3.0.1
